Add pre-commit configuration
I recognize that pre-commit strongly suggests to not use local
environments and instead rely on the environments created by the
pre-commit tool, but given that it's a stupid-ass suggestion, I've
elected to ignore it.

In all seriousness, we already have a virtualenv with the right packages
installed, which we need for our code editors. I do not want to deal
with having to sync those versions (yes, I know there is a tool for
that, but that makes it worse).