FileTree: resolve path relative to source dir
Currently, the path is resolved relative to the role directory, which is
undesirable. We want to be able to put a file tree anywhere in the role
directory and only copy the structure of everything below it.

For example, FileTree("foo/bar").replicates_self_to("/some/path") would
create a /some/path/foo/bar structure, while we would expect it to be
/some/path/bar.