Fix elevated command
This changes the sudo command to use --preserve-env flag, since sudo
resets the environment by default, so sudo env is essentially a no-op.

We also explicitly use the same executable for the non-elevated
execution, because even with --preserve-env, sudo resets the PATH
variable (among others) as a security measure.

Note that regardless of whether you run a Python script as 'python
script.py' or './script.py', sys.argv[0] will be './script.py', so
adding the executable will never lead to the executable being specified
twice.